implement `ac_auto` option for connected ac power
While the most recent change  introduced the `ac_auto` mode and made it
possible to select a mode to be used when  booting on battery power, it
would probably be useful to have the mode for external power configura-
ble too, e. g. to enable a 'Intel GPU on battery power / Hybrid mode on
ac-power' setup.

This commit adds an according setting to  the config file, accepting as
value either `hybrid` or `nvidia`. Further, the `README.md` file is ex-
tended to document this new option.